+++ dailyWeb/src/main/java/cn/wenzhuo4657/dailyWeb/domain/Types/model/dto/TypeDto.java
@@ -2,22 +2,22 @@ package cn.wenzhuo4657.dailyWeb.domain.Types.model.dto;
 
 public class TypeDto {
 
-    private Integer id;
+    private Long id;
     private String name;
 
     public TypeDto() {
     }
 
-    public TypeDto(Integer id, String name) {
+    public TypeDto(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
