+++ dailyWeb/src/main/java/cn/wenzhuo4657/dailyWeb/domain/Types/repository/ITypesRepository.java
@@ -12,5 +12,5 @@ public interface ITypesRepository {
 
     List<TypeDto> getAllByUserId(Long userId);
 
-    List<Docs> getDocsIdByTypeId(Long userId, Integer typeId);
+    List<Docs> getDocsIdByTypeId(Long userId, Long typeId);
 }
