+++ dailyWeb/src/main/java/cn/wenzhuo4657/dailyWeb/domain/tell/service/ITellService.java
@@ -16,7 +16,7 @@ public interface ITellService {
      * @param message 通知消息
      *
      */
-    public NotifierResult sendNotify(Integer notifyId, String type, NotifierMessage message);
+    public NotifierResult sendNotify(Long notifyId, String type, NotifierMessage message);
 
 
     /**
